Estimate empirical p-values for \(Chi^2\) protein correlations.

Andersen et al. (2003) used a fixed \(Chi^2\) threshold of 0.05 to identify organelle-specific candidates. This function computes empirical p-values by permutation the markers relative intensities and computed null \(Chi^2\) values.

Usage

empPvalues(marker, corMatrix, n = 100, ...)

Arguments

marker

A numerics with markers relative intensities.

corMatrix

A matrix of nrow(corMatrix) protein relative intensities to be compares against the marker.

n

The number of iterations.

...

Additional parameters to be passed to chi2.

Value

A numeric of length nrow(corMatrix).

Author

Laurent Gatto <lg390@cam.ac.uk>

See also

References

Andersen, J. S., Wilkinson, C. J., Mayor, T., Mortensen, P. et al., Proteomic characterization of the human centrosome by protein correlation profiling. Nature 2003, 426, 570 - 574.

Examples

set.seed(1)
mrk <- rnorm(6, 5, 1)
prot <- rbind(matrix(rnorm(120, 5, 1), ncol = 6),
              mrk + rnorm(6))
mrk <- mrk/sum(mrk)
prot <- prot/rowSums(prot)
empPvalues(mrk, prot)
#>  [1] 0.92 0.43 0.62 0.23 0.49 0.49 0.42 0.19 0.85 0.59 0.83 0.44 0.78 0.83 0.92
#> [16] 0.90 0.34 0.62 0.18 0.32 0.00