Update a feature variable

This function replaces a string or regular expression in a feature variable using the sub function.

fDataToUnknown(object, fcol = "markers", from = "^$", to = "unknown", ...)

Arguments

object
An instance of class MSnSet.

fcol
Feature variable to be modified. Default is "markers". If NULL, all feature variables will updated.

from
A character defining the string or regular expression of the pattern to be replaced. Default is the empty string, i.e. the regular expression "^$". See sub for details. If NA, then NA values are replaced by to.

to
A replacement for matched pattern. Default is "unknown". See sub for details.

...
Additional arguments passed to sub.

Value

An updated MSnSet.

Examples

library("pRolocdata")
data(dunkley2006)
getMarkers(dunkley2006, "markers")

#> organelleMarkers
#>      ER lumen   ER membrane         Golgi Mitochondrion       Plastid 
#>            14            45            28            55            20 
#>            PM      Ribosome           TGN       unknown       vacuole 
#>            46            19            13           428            21 

dunkley2006 <- fDataToUnknown(dunkley2006,
                              from = "unknown", to = "unassigned")
getMarkers(dunkley2006, "markers")

#> organelleMarkers
#>      ER lumen   ER membrane         Golgi Mitochondrion       Plastid 
#>            14            45            28            55            20 
#>            PM      Ribosome           TGN    unassigned       vacuole 
#>            46            19            13           428            21