Removes class/annotation information from a matrix of candidate markers that appear in the fData.

Removes annotation information that contain more that a certain number/percentage of proteins

filterMaxMarkers(object, n, p = 0.2, fcol = "GOAnnotations",
  verbose = TRUE)

Arguments

object
An instance of class MSnSet.

n
Maximum number of proteins allowed per class/information term.

p
Maximum percentage of proteins per column. Default is 0.2 i.e. remove columns that have information for greater than 20 of the total number of proteins in the dataset (note: this is useful for example, if information is GO terms, for removing very general and uninformative terms).

fcol
The name of the matrix of marker information. Default is GOAnnotations.

verbose
Number of marker candidates retained after filtering.

Value

An updated MSnSet

See also