Retrieve GO terms for feature names

The function pulls the gene ontology (GO) terms for a set of feature names.

getGOFromFeatures(id, namespace = "cellular_component",
  evidence = NULL, params = NULL, verbose = FALSE, nmax = 500)

Arguments

id

An character with feature names to be pulled from biomart. If and MSnSet is provided, then featureNames(id) is used.

namespace

The GO namespace. One of biological_process, cellular_component (default) or molecular_function.

evidence

The GO evidence code. See showGOEvidenceCodes for details. If NULL (default), no filtering based on the evidence code is performed.

params

An instance of class "AnnotationParams".

verbose

A logical defining verbosity of the function. Default is FALSE.

nmax

As described in https://support.bioconductor.org/p/86358/, the Biomart result can be unreliable for large queries. This argument splits the input in chunks of length nmax (default is 500). If set to NULL, the query is performed in full.

Value

A data.frame with relevant GO terms.

Author

Laurent Gatto

Examples

library(pRolocdata)
data(dunkley2006)
data(dunkley2006params)
dunkley2006params
#> Object of class "AnnotationParams"
#>  Using the 'plants_mart' BioMart database
#>  Using the 'athaliana_eg_gene' dataset
#>  Using 'tair_locus' as filter
#>  Created on Sat May  1 00:06:25 2021
fn <- featureNames(dunkley2006)[1:5]
getGOFromFeatures(fn, params = dunkley2006params)
#>    tair_locus      go_id     namespace_1003
#> 3   AT1G09210 GO:0005783 cellular_component
#> 8   AT1G09210 GO:0005788 cellular_component
#> 9   AT1G09210 GO:0005789 cellular_component
#> 13  AT1G09210 GO:0005783 cellular_component
#> 14  AT1G09210 GO:0005576 cellular_component
#> 15  AT1G09210 GO:0005739 cellular_component
#> 16  AT1G09210 GO:0099503 cellular_component
#> 18  AT1G09210 GO:0000325 cellular_component
#> 21  AT1G21750 GO:0005783 cellular_component
#> 22  AT1G21750 GO:0005634 cellular_component
#> 25  AT1G21750 GO:0005773 cellular_component
#> 28  AT1G21750 GO:0005788 cellular_component
#> 29  AT1G21750 GO:0005783 cellular_component
#> 30  AT1G21750 GO:0005783 cellular_component
#> 34  AT1G21750 GO:0099503 cellular_component
#> 35  AT1G21750 GO:0009579 cellular_component
#> 37  AT1G21750 GO:0000325 cellular_component
#> 40  AT1G21750 GO:0000326 cellular_component
#> 41  AT1G21750 GO:0000327 cellular_component
#> 44  AT1G51760 GO:0005783 cellular_component
#> 45  AT1G51760 GO:0005634 cellular_component
#> 46  AT1G51760 GO:0005788 cellular_component
#> 48  AT1G51760 GO:0005783 cellular_component
#> 56  AT1G56340 GO:0005783 cellular_component
#> 59  AT1G56340 GO:0005788 cellular_component
#> 61  AT1G56340 GO:0005634 cellular_component
#> 64  AT1G56340 GO:0005789 cellular_component
#> 67  AT1G56340 GO:0009506 cellular_component
#> 68  AT1G56340 GO:0005783 cellular_component
#> 69  AT1G56340 GO:0005739 cellular_component
#> 70  AT1G56340 GO:0099503 cellular_component
#> 72  AT1G56340 GO:0000325 cellular_component
#> 74  AT2G32920 GO:0005783 cellular_component
#> 77  AT2G32920 GO:0005788 cellular_component
#> 78  AT2G32920 GO:0005783 cellular_component
#> 80  AT2G32920 GO:0005794 cellular_component
#> 83  AT2G32920 GO:0000325 cellular_component
#> 84  AT2G32920 GO:0005788 cellular_component
#>                                       name_1006 go_linkage_type
#> 3                         endoplasmic reticulum             IEA
#> 8                   endoplasmic reticulum lumen             IEA
#> 9                endoplasmic reticulum membrane             IBA
#> 13                        endoplasmic reticulum             HDA
#> 14                         extracellular region             HDA
#> 15                                mitochondrion             HDA
#> 16                            secretory vesicle             HDA
#> 18                           plant-type vacuole             HDA
#> 21                        endoplasmic reticulum             IEA
#> 22                                      nucleus             HDA
#> 25                                      vacuole             IEA
#> 28                  endoplasmic reticulum lumen             IEA
#> 29                        endoplasmic reticulum             IBA
#> 30                        endoplasmic reticulum             HDA
#> 34                            secretory vesicle             HDA
#> 35                                    thylakoid             HDA
#> 37                           plant-type vacuole             HDA
#> 40                      protein storage vacuole             IDA
#> 41 lytic vacuole within protein storage vacuole             IDA
#> 44                        endoplasmic reticulum             IEA
#> 45                                      nucleus             HDA
#> 46                  endoplasmic reticulum lumen             IEA
#> 48                        endoplasmic reticulum             HDA
#> 56                        endoplasmic reticulum             IEA
#> 59                  endoplasmic reticulum lumen             IEA
#> 61                                      nucleus             HDA
#> 64               endoplasmic reticulum membrane             IBA
#> 67                                  plasmodesma             HDA
#> 68                        endoplasmic reticulum             HDA
#> 69                                mitochondrion             HDA
#> 70                            secretory vesicle             HDA
#> 72                           plant-type vacuole             HDA
#> 74                        endoplasmic reticulum             IEA
#> 77                  endoplasmic reticulum lumen             IEA
#> 78                        endoplasmic reticulum             HDA
#> 80                              Golgi apparatus             HDA
#> 83                           plant-type vacuole             HDA
#> 84                  endoplasmic reticulum lumen             IBA