Returns the organelle classes in an 'MSnSet'

Convenience accessor to the organelle classes in an 'MSnSet'. This function returns the organelle classes of an MSnSet instance. As a side effect, it prints out the classes.

Usage

getMarkerClasses(object, fcol = "markers", ...)

Arguments

object

An instance of class "MSnSet".

fcol

The name of the markers column in the featureData slot. Default is markers.

...

Additional parameters passed to sort from the base package.

Value

A character vector of the organelle classes in the data.

See also

Author

Lisa Breckels and Laurent Gatto

Examples

library("pRolocdata")
data(dunkley2006)
organelles <- getMarkerClasses(dunkley2006)
## same if markers encoded as a matrix
dunkley2006 <- mrkVecToMat(dunkley2006, mfcol = "Markers")
organelles2 <- getMarkerClasses(dunkley2006, fcol = "Markers")
stopifnot(all.equal(organelles, organelles2))