A complete LOPIMS pipeline

The function processes MSe data using the synergise function of the synapter package and combines resulting Synapter instances into one "MSnSet" and organelle marker data is added as a feature-level annotation variable.

lopims(hdmsedir = "HDMSE", msedir = "MSE", pep3ddir = "pep3D",
  fastafile, markerfile, mfdr = 0.025, ...)

Arguments

hdmsedir
A character identifying the directory containing the HDMSe final peptide files. Default is HDMSe.

msedir
A character identifying the directory containing the MSe final peptide files. Default is MSe.

pep3ddir
A character identifying the directory containing the MSe pep 3D files. Default is pep3D.

fastafile
A character identifying the protein fasta database. Default is to use the fasta file in the current directory. If several such files exist, the function reports an error.

markerfile
A character identifying the marker file (see details for format). Default is to use a csv file starting with marker in the current directory. If several such files exist, the function reports an error.

mfdr
The master FDR value. Default is 0.025.

...
Additional paramters passed to synergise.

Value

An instance of class "MSnSet" with protein level quantitation and respective organelle markers.

Details

The LOPIMS pipeline is composed of 5 steps:

The HDMSe final peptide files are used to compute false discovery rates uppon all possible combinations of HDMSe final peptides files and the best combination smaller or equal to mfdr is chosen. See estimateMasterFdr for details. The corresponding master run is then created as descibed in makeMaster. (function lopims1)

Each MSe/pep3D pair is processed using the HDMSe master file using synergise. (function lopims2)

The respective peptide-level synergise output objects are converted and combined into an single "MSnSet" instance. (function lopims3)

Protein-level quantitation is inferred as follows. For each protein, a reference sample/fraction is chosen based on the number of missing values (NA). If several samples have a same minimal number of NAs, ties are broken using the sum of counts. The peptides that do not display any missing values for each (frac_i, frac_ref) pair are summed and the ratio is reported (see pRoloc:::refNormMeanOfNonNAPepSum for details). (function lopims4)

The markers defined in the markerfile are collated as feature meta-data in the markers variable. See addMarkers for details. (function lopims5)

Intermediate synergise reports as well as resulting objects are stored in a LOPIMS_pipeline directory. For details, please refer to the synapter vignette and reference papers.

References

Improving qualitative and quantitative performance for MSE-based label free proteomics N.J. Bond, P.V. Shliaha, K.S. Lilley and L. Gatto Journal of Proteome Research, 2013;12(6):2340-53. PMID: 23510225.

The Effects of Travelling Wave Ion Mobility Separation on Data Independent Acquisition in Proteomics Studies P.V. Shliaha, N.J. Bond, L. Gatto and K.S. Lilley Journal of Proteome Research, 2013;12(6):2323-39. PMID: 23514362.

MSnbase-an R/Bioconductor package for isobaric tagged mass spectrometry data visualization, processing and quantitation. L. Gatto and KS. Lilley. Bioinformatics. 2012 Jan 15;28(2):288-9. doi: 10.1093/bioinformatics/btr645. Epub 2011 Nov 22. PubMed PMID: 22113085.

Author

Laurent Gatto