Number of outlier at each iteration of MCMC

Helper function to get the number of outlier at each MCMC iteration.

Helper function to get mean component allocation at each MCMC iteration.

Helper function to get mean probability of belonging to outlier at each iteration.

Wrapper for the geweke diagnostics from coda package also return p-values.

Helper function to pool chains together after processing

Helper function to burn n iterations from the front of the chains

Helper function to subsample the chains, known informally as thinning.

Produces a violin plot with the protein posterior probabilities distributions for all organelles.

Usage

mcmc_get_outliers(x)

mcmc_get_meanComponent(x)

mcmc_get_meanoutliersProb(x)

geweke_test(k)

mcmc_pool_chains(param)

mcmc_burn_chains(x, n = 50)

mcmc_thin_chains(x, freq = 5)

# S4 method for class 'MCMCParams,character'
plot(x, y, ...)

Arguments

x

Object of class MCMCParams

k

A list of coda::mcmc objects, as returned by mcmc_get_outliers, mcmc_get_meanComponent and mcmc_get_meanoutliersProb.

param

An object of class MCMCParams.

n

integer(1) defining number of iterations to burn. The default is 50

freq

Thinning frequency. The function retains every `freq`th iteration and is an `integer(1)`. The default thinning frequency is `5`.

y

A `character(1)` with a protein name.

...

Currently ignored.

Value

A list of length length(x).

A list of length length(x).

A list of length length(x).

A matrix with the test z- and p-values for each chain.

A pooled MCMCParams object.

An updated MCMCParams object.

A thinned `MCMCParams` object.

A ggplot2 object.

Author

Laurent Gatto