Creates a reduced marker variable

This function updates an MSnSet instances and sets markers class to unknown if there are less than n instances.

minMarkers(object, n = 10, fcol = "markers")

Arguments

object

An instance of class "MSnSet".

n

Minumum of marker instances per class.

fcol

The name of the markers column in the featureData slot. Default is markers.

Value

An instance of class "MSnSet" with a new feature variables, named after the original fcol variable and the n value.

See also

getPredictions to filter based on classification scores.

Examples

library(pRolocdata)
data(dunkley2006)
d2 <- minMarkers(dunkley2006, 20)
getMarkers(dunkley2006)

#> organelleMarkers
#>      ER lumen   ER membrane         Golgi Mitochondrion       Plastid 
#>            14            45            28            55            20 
#>            PM      Ribosome           TGN       unknown       vacuole 
#>            46            19            13           428            21 

getMarkers(d2, fcol = "markers20")

#> organelleMarkers
#>   ER membrane         Golgi Mitochondrion       Plastid            PM 
#>            45            28            55            20            46 
#>       unknown       vacuole 
#>           474            21