Model calibration plots

Model calibration model with posterior z-scores and posterior shrinkage

mixing_posterior_check(object, params, priors, fcol = "markers")

Arguments

object
A valid object of class MSnset

params
A valid object of class MCMCParams that has been processed and checked for convergence

priors
The prior that were used in the model

fcol
The columns of the feature data which contain the marker data.

Value

Used for side effect of producing plot. Invisibily returns an ggplot object that can be further manipulated

Author

Oliver M. Crook <omc25@cam.ac.uk>

Examples

if (FALSE) {
library("pRoloc")
data("tan2009r1")

tanres <- tagmMcmcTrain(object = tan2009r1)
tanres <- tagmMcmcProcess(tanres)
tan2009r1 <- tagmMcmcPredict(object = tan2009r1, params = tanres, probJoint = TRUE)
myparams <- chains(e14Tagm_converged_pooled)[[1]]
myparams2 <- chains(mcmc_pool_chains(tanres))[[1]]
priors <- tanres@priors
pRoloc:::mixing_posterior_check(object = tan2009r1, params = myparams2, priors = priors)
}