Uncertainty plot organelle means

Produces a pca plot with uncertainty in organelle means projected onto the PCA plot with contours.

Usage

nicheMeans2D(
  object,
  params,
  priors,
  dims = c(1, 2),
  fcol = "markers",
  aspect = 0.5
)

Arguments

object

A valid object of class MSnset

params

A valid object of class MCMCParams that has been processed and checked for convergence

priors

The prior that were used in the model

dims

The PCA dimension in which to project he data, default is c(1,2)

fcol

The columns of the feature data which contain the marker data.

aspect

A argument to change the plotting aspect of the PCA

Value

Used for side effect of producing plot. Invisibily returns an ggplot object that can be further manipulated

Author

Oliver M. Crook <omc25@cam.ac.uk>

Examples

if (FALSE) { # \dontrun{
library("pRolocdata")
data("tan2009r1")

tanres <- tagmMcmcTrain(object = tan2009r1)
tanres <- tagmMcmcProcess(tanres)
tan2009r1 <- tagmMcmcPredict(object = tan2009r1, params = tanres, probJoint = TRUE)
myparams <- chains(e14Tagm_converged_pooled)[[1]]
myparams2 <- chains(mcmc_pool_chains(tanres))[[1]]
priors <- tanres@priors
pRoloc:::nicheMeans2D(object = tan2009r1, params = myparams2, priors = priors)
} # }