This package implements pattern recognition techniques on quantitiative mass spectrometry data to infer protein sub-cellular localisation.

Details

More details about the package a provided in the following vignettes

pRoloc-ml

An overview of the machine learning techniques available in the pRoloc package.

pRoloc-tutorial

The main pRoloc tutorial, providing a hands-on introduction to the package, including data requirements, visualisation, clustering, classification and the application of semi-supervised machine learning.

pRoloc-transfer-learning

Description of a transfer learning algorithm for spatial proteomics.

HUPO_2011_poster

HUPO 2011 poster: pRoloc - A unifying bioinformatics framework for organelle proteomics.

HUPO_2014_poster

HUPO 2014 poster: A state-of-the-art machine learning pipeline for the analysis of spatial proteomics data.

If you have questions, want to rebort a bug or share suggestions, please file an issue at https://github.com/lgatto/MSnbase/issues, contact me directly or ask a question on the Bioconductor support forum https://support.bioconductor.org/.

References

Gatto L, Breckels LM, Wieczorek S, Burger T, Lilley KS. Mass-spectrometry-based spatial proteomics data analysis using pRoloc and pRolocdata. Bioinformatics. 2014 May 1;30(9):1322-4. doi:10.1093/bioinformatics/btu013. Epub 2014 Jan 11. PubMed PMID: 24413670; PubMed Central PMCID: PMC3998135.

Breckels LM, Gatto L, Christoforou A, Groen AJ, Lilley KS, Trotter MW. The effect of organelle discovery upon sub-cellular protein localisation. J Proteomics. 2013 Mar 21. doi:pii: S1874-3919(13)00094-8. 10.1016/j.jprot.2013.02.019. PubMed PMID: 23523639.

Gatto L., Breckels L.M., Burger T, Nightingale D.J.H., Groen A.J., Campbell C., Mulvey C.M., Christoforou A., Ferro M., Lilley K.S. 'A foundation for reliable spatial proteomics data analysis' Mol Cell Proteomics. 2014 May 20.

See also

The underlying infrastructure to store and manipulate the quantitative data is implemented in the MSnbase package. See MSnbase to get started.