A function to plot probabiltiy ellipses on marker PCA plots to visualise and assess TAGM models.

Note that when running PCA, this function does not scale the data (centring is performed), as opposed to [plot2D()]. Only marker proteins are displayed; the protein of unknown location, that are not used to estimate the MAP parameters, are filtered out.

Usage

plotEllipse(object, params, dims = c(1, 2), method = "MAP", ...)

Arguments

object

An [`MSnbase::MSnset`] containing quantitative spatial proteomics data.

params

An [`MAPParams`] with the TAGM-MAP parameters, as generated by `tagmMapTrain`.

dims

A `numeric(2)` with the principal components along which to project the data. Default is `c(1, 2)`.

method

The method used. Currently `"MAP"` only.

...

Additional parameters passed to [plot2D()].

Value

A PCA plot of the marker data with probability ellipises. The outer ellipse contains 99 probability whilst the middle and inner ellipses contain 95 and 90 clusters are represented by black circumpunct (circled dot).

See also