Extract a stratified sample of an

This function extracts a stratified sample of an MSnSet.

sampleMSnSet(object, fcol = "markers", size = 0.2, seed)

Arguments

object

An instance of class MSnSet

fcol

The feature meta-data column name containing the marker (vector or matrix) definitions on which the MSnSet will be stratified. Default is markers.

size

The size of the stratified sample to be extracted. Default is 0.2 (20 percent).

seed

The optional random number generator seed.

Value

A stratified sample (according to the defined fcol) which is an instance of class "MSnSet".

See also

testMSnSet unknownMSnSet markerMSnSet. See markers for details about markers encoding.

Examples

library(pRolocdata)
data(tan2009r1)
dim(tan2009r1)

#> [1] 888   4

smp <- sampleMSnSet(tan2009r1, fcol = "markers")

#> Warning: New sample contains classes with < 6 markers

dim(smp)

#> [1] 182   4

getMarkers(tan2009r1)

#> organelleMarkers
#>  Cytoskeleton            ER         Golgi      Lysosome mitochondrion 
#>             7            28            13             8            29 
#>       Nucleus    Peroxisome            PM    Proteasome  Ribosome 40S 
#>            21             4            34            15            20 
#>  Ribosome 60S       unknown 
#>            32           677 

getMarkers(smp)

#> organelleMarkers
#>  Cytoskeleton            ER         Golgi      Lysosome mitochondrion 
#>             2             6             3             2             6 
#>       Nucleus    Peroxisome            PM    Proteasome  Ribosome 40S 
#>             5             1             7             3             4 
#>  Ribosome 60S       unknown 
#>             7           136