GO Evidence Codes

This function prints a textual description of the Gene Ontology evidence codes.

showGOEvidenceCodes()

getGOEvidenceCodes()

Value

These functions are used for their side effects of printing evidence codes and their description.

Examples

showGOEvidenceCodes()

#> GO Term Evidence Code
#>  Experimental Evidence Codes
#>   EXP: Inferred from Experiment
#>    IDA: Inferred from Direct Assay
#>    IPI: Inferred from Physical Interaction
#>    IMP: Inferred from Mutant Phenotype
#>    IGI: Inferred from Genetic Interaction
#>    IEP: Inferred from Expression Pattern
#>  Computational Analysis Evidence Codes
#>   ISS: Inferred from Sequence or Structural Similarity
#>    ISO: Inferred from Sequence Orthology
#>    ISA: Inferred from Sequence Alignment
#>    ISM: Inferred from Sequence Model
#>    IGC: Inferred from Genomic Context
#>    IBA: Inferred from Biological aspect of Ancestor
#>    IBD: Inferred from Biological aspect of Descendant
#>    IKR: Inferred from Key Residues
#>    IRD: Inferred from Rapid Divergence
#>    RCA: inferred from Reviewed Computational Analysis
#>  Author Statement Evidence Codes
#>    TAS: Traceable Author Statement
#>    NAS: Non-traceable Author Statement
#>  Curator Statement Evidence Codes
#>    IC: Inferred by Curator
#>    ND: No biological Data available
#>  Automatically-assigned Evidence Codes
#>    IEA: Inferred from Electronic Annotation
#>  Obsolete Evidence Codes
#>    NR: Not Recorded

getGOEvidenceCodes()

#>  [1] "EXP" "IDA" "IPI" "IMP" "IGI" "IEP" "ISS" "ISO" "ISA" "ISM" "IGC" "IBA"
#> [13] "IBD" "IKR" "IRD" "RCA" "TAS" "NAS" "IC"  "ND"  "IEA" "NR"