Instrastructure to store and process MCMC results

The MCMCParams infrastructure is used to store and process Marchov chain Monte Carlo results for the T-Augmented Gaussian Mixture model (TAGM) from Crook et al. (2018).

chains(object)

# S4 method for MCMCParams
show(object)

# S4 method for ComponentParam
show(object)

# S4 method for MCMCChain
show(object)

# S4 method for MCMCChains
length(x)

# S4 method for MCMCParams
length(x)

# S4 method for MCMCChains,ANY,ANY
[[(x, i, j = "missing", drop = "missing")

# S4 method for MCMCParams,ANY,ANY
[[(x, i, j = "missing", drop = "missing")

# S4 method for MCMCChains,ANY,ANY,ANY
[(x, i, j = "missing", drop = "missing")

# S4 method for MCMCParams,ANY,ANY,ANY
[(x, i, j = "missing", drop = "missing")

# S4 method for MCMCChains
show(object)

Arguments

object
An instance of appropriate class.

x
Object to be subset.

i
An integer(). Should be of length 1 for [[.

j
Missing.

drop
Missing.

Details

Objects of the MCMCParams class are created with the tagmMcmcTrain() function. These objects store the priors of the generative TAGM model and the results of the MCMC chains, which themselves are stored as an instance of class MCMCChains and can be accessed with the chains() function. A summary of the MCMC chains (or class MCMCSummary) can be further computed with the tagmMcmcProcess() function.

See the pRoloc-bayesian vignette for examples.

Slots

chains

list() containing the individual full MCMC chain results in an MCMCChains instance. Each element must be a valid MCMCChain instance.

posteriorEstimates

A data.frame documenting the prosterior priors in an MCMCSummary instance. It contains N rows and columns tagm.allocation, tagm.probability, tagm.outlier, tagm.probability.lowerquantile, tagm.probability.upperquantile and tagm.mean.shannon.

diagnostics

A matrix of dimensions 1 by 2 containing the MCMCSummary diagnostics.

tagm.joint

A matrix of dimensions N by K storing the joint probability in an MCMCSummary instance.

method

character(1) describing the method in the MCMCParams object.

chains

Object of class MCMCChains containing the full MCMC chain results stored in the MCMCParams object.

priors

list()

summary

Object of class MCMCSummary the summarised MCMC results available in the MCMCParams instance.

n

integer(1) indicating the number of MCMC interactions. Stored in an MCMCChain instance.

K

integer(1) indicating the number of components. Stored in an MCMCChain instance.

N

integer(1) indicating the number of proteins. Stored in an MCMCChain instance.

Component

matrix(N, n) component allocation results of an MCMCChain instance.

ComponentProb

matrix(N, n, K) component allocation probabilities of an MCMCChain instance.

Outlier

matrix(N, n) outlier allocation results.

OutlierProb

matrix(N, n, 2) outlier allocation probabilities of an MCMCChain instance.

See also