Check feature names overlap

Checks the marker and unknown feature overlap of two MSnSet instances.

Usage

checkFeatureNamesOverlap(x, y, fcolx = "markers", fcoly, verbose = TRUE)

Arguments

x

An MSnSet instance.

y

An MSnSet instance.

fcolx

The feature variable to separate unknown (fData(y)$coly == "unknown") from the marker features in the x object.

fcoly

As fcolx, for the y object. If missing, the value of fcolx is used.

verbose

If TRUE (default), the overlap is printed out on the console.

Value

Invisibly returns a named list of common markers, unique x markers, unique y markers in, common unknowns, unique x unknowns and unique y unknowns.

Author

Laurent Gatto

Examples

library("pRolocdata")
data(andy2011)
data(andy2011goCC)
checkFeatureNamesOverlap(andy2011, andy2011goCC)
#> Common markers:  404 
#> Unique x markers:  0 
#> Unique y markers:  0 
#> Common unkowns:  967 
#> Unique x unknowns:  0 
#> Unique y unknowns:  0 
featureNames(andy2011goCC)[1] <- "ABC"
res <- checkFeatureNamesOverlap(andy2011, andy2011goCC)
#> Common markers:  403 
#> Unique x markers:  1 
#> Unique y markers:  1 
#> Common unkowns:  967 
#> Unique x unknowns:  0 
#> Unique y unknowns:  0 
res$markersX
#> [1] "O00767"
res$markersY
#> [1] "ABC"