Compare a feature variable overlap

Extracts qualitative feature variables from two MSnSet instances and compares with a contingency table.

checkFvarOverlap(x, y, fcolx = "markers", fcoly, verbose = TRUE)

Arguments

x
An MSnSet instance.

y
An MSnSet instance.

fcolx
The feature variable to separate unknown (fData(y)$coly == "unknown") from the marker features in the x object.

fcoly
As fcolx, for the y object. If missing, the value of fcolx is used.

verbose
If TRUE (default), the contingency table of the the feature variables is printed out.

Value

Invisibly returns a named list with the values of the diagonal, upper and lower triangles of the contingency table.

Author

Laurent Gatto

Examples

library("pRolocdata")
data(dunkley2006)
res <- checkFvarOverlap(dunkley2006, dunkley2006,
                        "markers", "markers.orig")

#>                my
#> mx               ER Golgi mit/plastid  PM unknown vacuole
#>   ER lumen       11     0           0   0       3       0
#>   ER membrane    21     0           0   0      24       0
#>   Golgi           0    27           0   0       1       0
#>   Mitochondrion   0     0          13   0      42       0
#>   Plastid         0     0           7   0      13       0
#>   PM              0     0           0  28      18       0
#>   Ribosome        0     0           0   0      19       0
#>   TGN             0     0           0   0      13       0
#>   unknown        17     0           6   0     405       0
#>   vacuole         0     0           0   0       9      12

str(res)

#> List of 3
#>  $ matches         : int [1:6] 11 0 0 0 13 0
#>  $ lower.mismatches: int [1:39] 21 0 0 0 0 0 0 17 0 27 ...
#>  $ upper.mismatches: int [1:39] 21 0 0 0 0 0 0 17 0 27 ...