The PCP 'chi square' method

In the original protein correlation profiling (PCP), Andersen et al. use the peptide normalised profiles along gradient fractions and compared them with the reference profiles (or set of profiles) by computing \(Chi^2\) values, \(\frac{\sum (x_i - x_p)^2}{x_p}\), where \(x_i\) is the normalised value of the peptide in fraction i and \(x_p\) is the value of the marker (from Wiese et al., 2007). The protein \(Chi^2\) is then computed as the median of the peptide \(Chi^2\) values. Peptides and proteins with similar profiles to the markers will have small \(Chi^2\) values.

The chi2 methods implement this idea and compute such Chi^2 values for sets of proteins.

Methods

signature(x = "matrix", y = "matrix", method = "character", fun = "NULL", na.rm = "logical")

Compute nrow(x) times nrow(y) \(Chi^2\) values, for each x, y feature pair. Method is one of "Andersen2003" or "Wiese2007"; the former (default) computed the \(Chi^2\) as sum(y-x)^2/length(x), while the latter uses sum((y-x)^2/x). na.rm defines if missing values (NA and NaN) should be removed prior to summation. fun defines how to summarise the \(Chi^2\) values; default, NULL, does not combine the \(Chi^2\) values.

signature(x = "matrix", y = "numeric", method = "character", na.rm = "logical")

Computes nrow(x) \(Chi^2\) values, for all the \((x_i, y)\) pairs. See above for the other arguments.

signature(x = "numeric", y = "matrix", method = "character", na.rm = "logical")

Computes nrow(y) \(Chi^2\) values, for all the \((x, y_i)\) pairs. See above for the other arguments.

signature(x = "numeric", y = "numeric", method = "character", na.rm = "logical")

Computes the \(Chi^2\) value for the \((x, y)\) pairs. See above for the other arguments.

References

Andersen, J. S., Wilkinson, C. J., Mayor, T., Mortensen, P. et al., Proteomic characterization of the human centrosome by protein correlation profiling. Nature 2003, 426, 570 - 574.

Wiese, S., Gronemeyer, T., Ofman, R., Kunze, M. et al., Proteomics characterization of mouse kidney peroxisomes by tandem mass spectrometry and protein correlation profiling. Mol. Cell. Proteomics 2007, 6, 2045 - 2057.

See also

Author

Laurent Gatto <lg390@cam.ac.uk>

Examples

mrk <- rnorm(6)
prot <- matrix(rnorm(60), ncol = 6)
chi2(mrk, prot, method = "Andersen2003")
#>  [1] 0.5973439 1.6204618 3.1474448 3.8159721 1.9442386 2.8598635 0.9618657
#>  [8] 2.4775154 1.9620390 1.1837790
chi2(mrk, prot, method = "Wiese2007")
#>  [1]  -61.800616  -40.863913  -18.631162  -11.964521  -48.730051  -59.901604
#>  [7]   -9.380375 -183.807730  -78.874597  -35.056524

pepmark <- matrix(rnorm(18), ncol = 6)
pepprot <- matrix(rnorm(60), ncol = 6)
chi2(pepmark, pepprot)
#>            [,1]      [,2]      [,3]
#>  [1,] 2.4060471 0.7128013 0.9692423
#>  [2,] 1.7169127 1.2541456 3.1220592
#>  [3,] 1.7974853 0.5822429 2.0619888
#>  [4,] 3.0617084 1.0692017 2.0215062
#>  [5,] 1.7684921 0.5954362 0.8037104
#>  [6,] 1.3206710 0.6904117 2.4765364
#>  [7,] 3.4272234 1.5089956 3.5153397
#>  [8,] 0.9959838 1.0816578 3.1640995
#>  [9,] 3.4955556 1.7019460 2.3328391
#> [10,] 2.4304560 0.4566767 1.6098192
chi2(pepmark, pepprot, fun = sum)
#>  [1] 4.088091 6.093117 4.441717 6.152416 3.167639 4.487619 8.451559 5.241741
#>  [9] 7.530341 4.496952