Retrieve GO terms for feature names

The function pulls the gene ontology (GO) terms for a set of feature names.

getGOFromFeatures(id, namespace = "cellular_component",
  evidence = NULL, params = NULL, verbose = FALSE, nmax = 500)

Arguments

id
An character with feature names to be pulled from biomart. If and MSnSet is provided, then featureNames(id) is used.

namespace
The GO namespace. One of biological_process, cellular_component (default) or molecular_function.

evidence
The GO evidence code. See showGOEvidenceCodes for details. If NULL (default), no filtering based on the evidence code is performed.

params
An instance of class "AnnotationParams".

verbose
A logical defining verbosity of the function. Default is FALSE.

nmax
As described in https://support.bioconductor.org/p/86358/, the Biomart result can be unreliable for large queries. This argument splits the input in chunks of length nmax (default is 500). If set to NULL, the query is performed in full.

Value

A data.frame with relevant GO terms.

Examples

library(pRolocdata)
data(dunkley2006)
data(dunkley2006params)
dunkley2006params

#> Object of class "AnnotationParams"
#>  Using the 'plants_mart' BioMart database
#>  Using the 'athaliana_eg_gene' dataset
#>  Using 'tair_locus' as filter
#>  Created on Thu Apr 20 19:19:37 2017

fn <- featureNames(dunkley2006)[1:5]
getGOFromFeatures(fn, params = dunkley2006params)

#>    tair_locus      go_id     namespace_1003
#> 1   AT1G09210 GO:0005788 cellular_component
#> 2   AT1G09210 GO:0005773 cellular_component
#> 5   AT1G09210 GO:0005783 cellular_component
#> 11  AT1G09210 GO:0005783 cellular_component
#> 12  AT1G09210 GO:0048046 cellular_component
#> 16  AT1G21750 GO:0000327 cellular_component
#> 19  AT1G21750 GO:0000326 cellular_component
#> 21  AT1G21750 GO:0005788 cellular_component
#> 22  AT1G21750 GO:0005623 cellular_component
#> 24  AT1G21750 GO:0005773 cellular_component
#> 25  AT1G21750 GO:0009505 cellular_component
#> 28  AT1G21750 GO:0005783 cellular_component
#> 32  AT1G21750 GO:0009579 cellular_component
#> 35  AT1G21750 GO:0005783 cellular_component
#> 38  AT1G21750 GO:0005773 cellular_component
#> 43  AT1G21750 GO:0005783 cellular_component
#> 44  AT1G51760 GO:0005788 cellular_component
#> 47  AT1G51760 GO:0005783 cellular_component
#> 49  AT1G51760 GO:0005783 cellular_component
#> 54  AT1G56340 GO:0005783 cellular_component
#> 58  AT1G56340 GO:0005788 cellular_component
#> 59  AT1G56340 GO:0009506 cellular_component
#> 64  AT1G56340 GO:0005774 cellular_component
#> 66  AT1G56340 GO:0005783 cellular_component
#> 67  AT1G56340 GO:0048046 cellular_component
#> 71  AT2G32920 GO:0005788 cellular_component
#> 72  AT2G32920 GO:0005623 cellular_component
#> 73  AT2G32920 GO:0005773 cellular_component
#> 75  AT2G32920 GO:0005783 cellular_component
#> 78  AT2G32920 GO:0005783 cellular_component
#>                                       name_1006 go_linkage_type
#> 1                   endoplasmic reticulum lumen             IEA
#> 2                                       vacuole             IDA
#> 5                         endoplasmic reticulum             IEA
#> 11                        endoplasmic reticulum             IDA
#> 12                                     apoplast             IDA
#> 16 lytic vacuole within protein storage vacuole             IDA
#> 19                      protein storage vacuole             IDA
#> 21                  endoplasmic reticulum lumen             IEA
#> 22                                         cell             IEA
#> 24                                      vacuole             IDA
#> 25                         plant-type cell wall             IDA
#> 28                        endoplasmic reticulum             IEA
#> 32                                    thylakoid             IDA
#> 35                        endoplasmic reticulum             IDA
#> 38                                      vacuole             IEA
#> 43                        endoplasmic reticulum             IBA
#> 44                  endoplasmic reticulum lumen             IEA
#> 47                        endoplasmic reticulum             IEA
#> 49                        endoplasmic reticulum             IDA
#> 54                        endoplasmic reticulum             IEA
#> 58                  endoplasmic reticulum lumen             IEA
#> 59                                  plasmodesma             IDA
#> 64                            vacuolar membrane             IDA
#> 66                        endoplasmic reticulum             IDA
#> 67                                     apoplast             IDA
#> 71                  endoplasmic reticulum lumen             IEA
#> 72                                         cell             IEA
#> 73                                      vacuole             IDA
#> 75                        endoplasmic reticulum             IEA
#> 78                        endoplasmic reticulum             IDA