Retrieve GO terms for feature names

The function pulls the gene ontology (GO) terms for a set of feature names.

getGOFromFeatures(id, namespace = "cellular_component", evidence = NULL,
  params = NULL, verbose = FALSE, nmax = 500)

Arguments

id
An character with feature names to be pulled from biomart. If and MSnSet is provided, then featureNames(id) is used.

namespace
The GO namespace. One of biological_process, cellular_component (default) or molecular_function.

evidence
The GO evidence code. See showGOEvidenceCodes for details. If NULL (default), no filtering based on the evidence code is performed.

params
An instance of class "AnnotationParams".

verbose
A logical defining verbosity of the function. Default is FALSE.

nmax
As described in https://support.bioconductor.org/p/86358/, the Biomart result can be unreliable for large queries. This argument splits the input in chunks of length nmax (default is 500). If set to NULL, the query is performed in full.

Value

A data.frame with relevant GO terms.

Examples

library(pRolocdata)
data(dunkley2006)
data(dunkley2006params)
dunkley2006params

#> Object of class "AnnotationParams"
#>  Using the 'plants_mart' BioMart database
#>  Using the 'athaliana_eg_gene' dataset
#>  Using 'tair_locus' as filter
#>  Created on Thu Apr 20 19:19:37 2017

fn <- featureNames(dunkley2006)[1:5]
getGOFromFeatures(fn, params = dunkley2006params)

#>    tair_locus      go_id     namespace_1003
#> 3   AT1G09210 GO:0005783 cellular_component
#> 4   AT1G09210 GO:0005739 cellular_component
#> 8   AT1G09210 GO:0005773 cellular_component
#> 10  AT1G09210 GO:0005783 cellular_component
#> 11  AT1G09210 GO:0048046 cellular_component
#> 13  AT1G09210 GO:0005788 cellular_component
#> 16  AT1G21750 GO:0005783 cellular_component
#> 17  AT1G21750 GO:0009507 cellular_component
#> 18  AT1G21750 GO:0016020 cellular_component
#> 20  AT1G21750 GO:0009505 cellular_component
#> 21  AT1G21750 GO:0009579 cellular_component
#> 23  AT1G21750 GO:0005773 cellular_component
#> 25  AT1G21750 GO:0005783 cellular_component
#> 31  AT1G21750 GO:0005773 cellular_component
#> 35  AT1G21750 GO:0005788 cellular_component
#> 37  AT1G21750 GO:0000326 cellular_component
#> 38  AT1G21750 GO:0000327 cellular_component
#> 42  AT1G51760 GO:0005783 cellular_component
#> 43  AT1G51760 GO:0005886 cellular_component
#> 44  AT1G51760 GO:0005783 cellular_component
#>                                       name_1006 go_linkage_type
#> 3                         endoplasmic reticulum             IEA
#> 4                                 mitochondrion             IDA
#> 8                                       vacuole             IDA
#> 10                        endoplasmic reticulum             IDA
#> 11                                     apoplast             IDA
#> 13                  endoplasmic reticulum lumen             IEA
#> 16                        endoplasmic reticulum             IEA
#> 17                                  chloroplast             IDA
#> 18                                     membrane             IDA
#> 20                         plant-type cell wall             IDA
#> 21                                    thylakoid             IDA
#> 23                                      vacuole             IDA
#> 25                        endoplasmic reticulum             IDA
#> 31                                      vacuole             IEA
#> 35                  endoplasmic reticulum lumen             IEA
#> 37                      protein storage vacuole             IDA
#> 38 lytic vacuole within protein storage vacuole             IDA
#> 42                        endoplasmic reticulum             IEA
#> 43                              plasma membrane             IDA
#> 44                        endoplasmic reticulum             IDA
#>  [ reached getOption("max.print") -- omitted 16 rows ]