Retrieve GO terms for feature names

The function pulls the gene ontology (GO) terms for a set of feature names.

Usage

getGOFromFeatures(
  id,
  namespace = "cellular_component",
  evidence = NULL,
  params = NULL,
  verbose = FALSE,
  nmax = 500
)

Arguments

id

An character with feature names to be pulled from biomart. If and MSnSet is provided, then featureNames(id) is used.

namespace

The GO namespace. One of biological_process, cellular_component (default) or molecular_function.

evidence

The GO evidence code. See showGOEvidenceCodes for details. If NULL (default), no filtering based on the evidence code is performed.

params

An instance of class "AnnotationParams".

verbose

A logical defining verbosity of the function. Default is FALSE.

nmax

As described in https://support.bioconductor.org/p/86358/, the Biomart result can be unreliable for large queries. This argument splits the input in chunks of length nmax (default is 500). If set to NULL, the query is performed in full.

Value

A data.frame with relevant GO terms.

Author

Laurent Gatto

Examples

library(pRolocdata)
data(dunkley2006)
data(dunkley2006params)
dunkley2006params
#> Object of class "AnnotationParams"
#>  Using the 'plants_mart' BioMart database
#>  Using the 'athaliana_eg_gene' dataset
#>  Using 'tair_locus' as filter
#>  Created on Tue Mar 12 07:25:12 2024
fn <- featureNames(dunkley2006)[1:5]
getGOFromFeatures(fn, params = dunkley2006params)
#>    tair_locus      go_id     namespace_1003                   name_1006
#> 3   AT1G21750 GO:0005783 cellular_component       endoplasmic reticulum
#> 5   AT1G21750 GO:0005788 cellular_component endoplasmic reticulum lumen
#> 10  AT1G56340 GO:0005783 cellular_component       endoplasmic reticulum
#> 13  AT1G56340 GO:0005788 cellular_component endoplasmic reticulum lumen
#>    go_linkage_type
#> 3              IEA
#> 5              IEA
#> 10             IEA
#> 13             IEA