Retrieve GO terms for feature names

The function pulls the gene ontology (GO) terms for a set of feature names.

getGOFromFeatures(id, namespace = "cellular_component",
  evidence = NULL, params = NULL, verbose = FALSE, nmax = 500)

Arguments

id
An character with feature names to be pulled from biomart. If and MSnSet is provided, then featureNames(id) is used.

namespace
The GO namespace. One of biological_process, cellular_component (default) or molecular_function.

evidence
The GO evidence code. See showGOEvidenceCodes for details. If NULL (default), no filtering based on the evidence code is performed.

params
An instance of class "AnnotationParams".

verbose
A logical defining verbosity of the function. Default is FALSE.

nmax
As described in https://support.bioconductor.org/p/86358/, the Biomart result can be unreliable for large queries. This argument splits the input in chunks of length nmax (default is 500). If set to NULL, the query is performed in full.

Value

A data.frame with relevant GO terms.

Author

Laurent Gatto

Examples

library(pRolocdata)
data(dunkley2006)
data(dunkley2006params)
dunkley2006params
#> Object of class "AnnotationParams"
#>  Using the 'plants_mart' BioMart database
#>  Using the 'athaliana_eg_gene' dataset
#>  Using 'tair_locus' as filter
#>  Created on Sat May  1 00:06:25 2021
fn <- featureNames(dunkley2006)[1:5]
getGOFromFeatures(fn, params = dunkley2006params)
#>    tair_locus      go_id     namespace_1003
#> 1   AT1G09210 GO:0005788 cellular_component
#> 4   AT1G09210 GO:0005783 cellular_component
#> 9   AT1G09210 GO:0005789 cellular_component
#> 12  AT1G09210 GO:0005783 cellular_component
#> 13  AT1G09210 GO:0005576 cellular_component
#> 14  AT1G09210 GO:0005739 cellular_component
#> 15  AT1G09210 GO:0099503 cellular_component
#> 18  AT1G09210 GO:0005773 cellular_component
#> 20  AT1G21750 GO:0005788 cellular_component
#> 22  AT1G21750 GO:0000327 cellular_component
#> 24  AT1G21750 GO:0005783 cellular_component
#> 25  AT1G21750 GO:0005634 cellular_component
#> 28  AT1G21750 GO:0005773 cellular_component
#> 31  AT1G21750 GO:0005783 cellular_component
#> 32  AT1G21750 GO:0009579 cellular_component
#> 33  AT1G21750 GO:0005783 cellular_component
#> 37  AT1G21750 GO:0000326 cellular_component
#> 40  AT1G21750 GO:0099503 cellular_component
#> 42  AT1G21750 GO:0005773 cellular_component
#> 44  AT1G51760 GO:0005788 cellular_component
#> 47  AT1G51760 GO:0005783 cellular_component
#> 48  AT1G51760 GO:0005634 cellular_component
#> 49  AT1G51760 GO:0005783 cellular_component
#> 52  AT1G51760 GO:0005783 cellular_component
#> 53  AT1G51760 GO:0005886 cellular_component
#> 57  AT1G56340 GO:0005788 cellular_component
#> 60  AT1G56340 GO:0005783 cellular_component
#> 65  AT1G56340 GO:0005634 cellular_component
#> 67  AT1G56340 GO:0005789 cellular_component
#> 70  AT1G56340 GO:0005783 cellular_component
#> 71  AT1G56340 GO:0005739 cellular_component
#> 72  AT1G56340 GO:0005774 cellular_component
#> 73  AT1G56340 GO:0099503 cellular_component
#> 76  AT1G56340 GO:0009506 cellular_component
#> 78  AT2G32920 GO:0005788 cellular_component
#> 80  AT2G32920 GO:0005783 cellular_component
#> 82  AT2G32920 GO:0005783 cellular_component
#> 83  AT2G32920 GO:0005886 cellular_component
#> 85  AT2G32920 GO:0005794 cellular_component
#> 87  AT2G32920 GO:0005773 cellular_component
#> 89  AT2G32920 GO:0005783 cellular_component
#>                                       name_1006 go_linkage_type
#> 1                   endoplasmic reticulum lumen             IEA
#> 4                         endoplasmic reticulum             IEA
#> 9                endoplasmic reticulum membrane             IBA
#> 12                        endoplasmic reticulum             IDA
#> 13                         extracellular region             HDA
#> 14                                mitochondrion             IDA
#> 15                            secretory vesicle             HDA
#> 18                                      vacuole             IDA
#> 20                  endoplasmic reticulum lumen             IEA
#> 22 lytic vacuole within protein storage vacuole             IDA
#> 24                        endoplasmic reticulum             IEA
#> 25                                      nucleus             HDA
#> 28                                      vacuole             IEA
#> 31                        endoplasmic reticulum             IBA
#> 32                                    thylakoid             IDA
#> 33                        endoplasmic reticulum             IDA
#> 37                      protein storage vacuole             IDA
#> 40                            secretory vesicle             HDA
#> 42                                      vacuole             IDA
#> 44                  endoplasmic reticulum lumen             IEA
#> 47                        endoplasmic reticulum             IEA
#> 48                                      nucleus             HDA
#> 49                        endoplasmic reticulum             IBA
#> 52                        endoplasmic reticulum             IDA
#> 53                              plasma membrane             IDA
#> 57                  endoplasmic reticulum lumen             IEA
#> 60                        endoplasmic reticulum             IEA
#> 65                                      nucleus             HDA
#> 67               endoplasmic reticulum membrane             IBA
#> 70                        endoplasmic reticulum             IDA
#> 71                                mitochondrion             IDA
#> 72                            vacuolar membrane             IDA
#> 73                            secretory vesicle             HDA
#> 76                                  plasmodesma             IDA
#> 78                  endoplasmic reticulum lumen             IEA
#> 80                        endoplasmic reticulum             IEA
#> 82                        endoplasmic reticulum             IDA
#> 83                              plasma membrane             IDA
#> 85                              Golgi apparatus             IDA
#> 87                                      vacuole             IDA
#> 89                        endoplasmic reticulum             HDA