Retrieve GO terms for feature names

The function pulls the gene ontology (GO) terms for a set of feature names.

getGOFromFeatures(id, namespace = "cellular_component", evidence = NULL,
  params = NULL, verbose = FALSE, nmax = 500)

Arguments

id

An character with feature names to be pulled from biomart. If and MSnSet is provided, then featureNames(id) is used.

namespace

The GO namespace. One of biological_process, cellular_component (default) or molecular_function.

evidence

The GO evidence code. See showGOEvidenceCodes for details. If NULL (default), no filtering based on the evidence code is performed.

params

An instance of class "AnnotationParams".

verbose

A logical defining verbosity of the function. Default is FALSE.

nmax

As described in https://support.bioconductor.org/p/86358/, the Biomart result can be unreliable for large queries. This argument splits the input in chunks of length nmax (default is 500). If set to NULL, the query is performed in full.

Value

A data.frame with relevant GO terms.

Examples

library(pRolocdata)
data(dunkley2006)
data(dunkley2006params)
dunkley2006params

#> Object of class "AnnotationParams"
#>  Using the 'plants_mart' BioMart database
#>  Using the 'athaliana_eg_gene' dataset
#>  Using 'tair_locus' as filter
#>  Created on Thu Apr 20 19:19:37 2017

fn <- featureNames(dunkley2006)[1:5]
getGOFromFeatures(fn, params = dunkley2006params)

#>    tair_locus      go_id     namespace_1003
#> 1   AT2G32920 GO:0005886 cellular_component
#> 2   AT2G32920 GO:0005794 cellular_component
#> 3   AT2G32920 GO:0005788 cellular_component
#> 4   AT2G32920 GO:0005783 cellular_component
#> 5   AT2G32920 GO:0005783 cellular_component
#> 6   AT2G32920 GO:0005773 cellular_component
#> 13  AT2G32920 GO:0005623 cellular_component
#> 15  AT1G21750 GO:0016020 cellular_component
#> 16  AT1G21750 GO:0009579 cellular_component
#> 17  AT1G21750 GO:0009507 cellular_component
#> 18  AT1G21750 GO:0009505 cellular_component
#> 19  AT1G21750 GO:0005788 cellular_component
#> 20  AT1G21750 GO:0005783 cellular_component
#> 21  AT1G21750 GO:0005783 cellular_component
#> 22  AT1G21750 GO:0005773 cellular_component
#> 23  AT1G21750 GO:0005773 cellular_component
#> 24  AT1G21750 GO:0000327 cellular_component
#> 25  AT1G21750 GO:0000326 cellular_component
#> 40  AT1G56340 GO:0005783 cellular_component
#> 44  AT1G56340 GO:0048046 cellular_component
#>                                       name_1006 go_linkage_type
#> 1                               plasma membrane             IDA
#> 2                               Golgi apparatus             IDA
#> 3                   endoplasmic reticulum lumen             IEA
#> 4                         endoplasmic reticulum             IDA
#> 5                         endoplasmic reticulum             IEA
#> 6                                       vacuole             IDA
#> 13                                         cell             IEA
#> 15                                     membrane             IDA
#> 16                                    thylakoid             IDA
#> 17                                  chloroplast             IDA
#> 18                         plant-type cell wall             IDA
#> 19                  endoplasmic reticulum lumen             IEA
#> 20                        endoplasmic reticulum             IDA
#> 21                        endoplasmic reticulum             IEA
#> 22                                      vacuole             IDA
#> 23                                      vacuole             IEA
#> 24 lytic vacuole within protein storage vacuole             IDA
#> 25                      protein storage vacuole             IDA
#> 40                        endoplasmic reticulum             IEA
#> 44                                     apoplast             IDA
#>  [ reached getOption("max.print") -- omitted 16 rows ]