Returns organelle-specific quantile scores

This function produces organelle-specific quantiles corresponding to the given classification scores.

orgQuants(object, fcol, scol, mcol = "markers", t, verbose = TRUE)

Arguments

object
An instance of class "MSnSet".

fcol
The name of the prediction column in the featureData slot.

scol
The name of the prediction score column in the featureData slot. If missing, created by pasting '.scores' after fcol.

mcol
The name of the column containing the training data in the featureData slot. Default is markers.

t
The quantile threshold.

verbose
If TRUE, the calculated threholds are printed.

Value

A named vector of organelle thresholds.

See also

Author

Lisa Breckels

Examples

library("pRolocdata")
data(dunkley2006)
res <- svmClassification(dunkley2006, fcol = "pd.markers",
                         sigma = 0.1, cost = 0.5)
## 50% top predictions per class
ts <- orgQuants(res, fcol = "svm", t = .5)

#>      ER lumen   ER membrane         Golgi Mitochondrion       Plastid 
#>     0.3193267     0.8325586     0.7755289     0.7476421     0.7714467 
#>            PM      Ribosome           TGN       vacuole 
#>     0.7292333     0.5101448     0.5378267     0.5608895 

getPredictions(res, fcol = "svm", t = ts)

#> ans
#>      ER lumen   ER membrane         Golgi Mitochondrion       Plastid 
#>            15           118            64            78            36 
#>            PM      Ribosome           TGN       unknown       vacuole 
#>            86            38            16           212            26 

#> MSnSet (storageMode: lockedEnvironment)
#> assayData: 689 features, 16 samples 
#>   element names: exprs 
#> protocolData: none
#> phenoData
#>   sampleNames: M1F1A M1F4A ... M2F11B (16 total)
#>   varLabels: membrane.prep fraction replicate
#>   varMetadata: labelDescription
#> featureData
#>   featureNames: AT1G09210 AT1G21750 ... AT4G39080 (689 total)
#>   fvarLabels: assigned evidence ... svm.pred (11 total)
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'
#>   pubMedIds: 16618929 
#> Annotation:  
#> - - - Processing information - - -
#> Loaded on Thu Jul 16 22:53:08 2015. 
#> Normalised to sum of intensities. 
#> Added markers from  'mrk' marker vector. Thu Jul 16 22:53:08 2015 
#> Performed svm prediction (cost=0.5 sigma=0.1) Sat May  1 11:59:16 2021 
#> Added svm predictions according to thresholds: ER lumen = 0.32, ER membrane = 0.83, Golgi = 0.78, Mitochondrion = 0.75, Plastid = 0.77, PM = 0.73, Ribosome = 0.51, TGN = 0.54, vacuole = 0.56 Sat May  1 11:59:16 2021 
#>  MSnbase version: 1.17.12