Organelle markers

This function retrieves a list of organelle markers or, if no species is provided, prints a description of available marker sets. The markers can be added to and MSnSet using the addMarkers function. Several marker version are provided (see Details for additional information).

Usage

pRolocmarkers(species, version = "2")

Arguments

species

character(1) defining the species of interest. For reference species markers, this is just the species e.g. "hsap". For published marker sets this is the species and author name e.g. "hsap_geladaki".

version

character(1) defining the marker version. Default is "2".

Value

Prints a description of the available marker lists if species is missing or a named character with organelle markers.

Details

Version 1 of the markers have been contributed by various members of the Cambridge Centre for Proteomics, in particular Dr Dan Nightingale for yeast, Dr Andy Christoforou and Dr Claire Mulvey for human, Dr Arnoud Groen for Arabodopsis and Dr Claire Mulvey for mouse. In addition, original (curated) markers from the pRolocdata datasets have been extracted (see pRolocdata for details and references). Curation involved verification of publicly available subcellular localisation annotation based on the curators knowledge of the organelles/proteins considered and tracing the original statement in the literature.

Version 2 of the markers (current default) have been updated by Charlotte Hutchings from the Cambridge Centre for Proteomics. Reference species marker sets are the same as those in version 1 with minor corrections and an updated naming system. Version 2 also contains additional marker sets from spatial proteomics publications. References for the source publications are provided below:

Geladaki, A., Britovsek, N.K., Breckels, L.M., Smith, T.S., Vennard, O.L., Mulvey, C.M., Crook, O.M., Gatto, L. and Lilley, K.S. (2019) Combining LOPIT with differential ultracentrifugation for high-resolution spatial proteomics. Nature Communications. 10 (1). doi:10.1038/s41467-018-08191-w

Christopher, J.A., Breckels, L.M., Crook, O.M., Vazquez–Chantada, M., Barratt, D. and Lilley, K.S. (2024) Global proteomics indicates subcellular-specific anti-ferroptotic responses to ionizing radiation.p.2024.09.12.611851. doi:10.1101/2024.09.12.611851

Itzhak, D.N., Tyanova, S., Cox, J. and Borner, G.H. (2016) Global, quantitative and dynamic mapping of protein subcellular localization. eLife. 5. doi:10.7554/elife.16950

Villanueva, E., Smith, T., Pizzinga, M., Elzek, M., Queiroz, R.M.L., Harvey, R.F., Breckels, L.M., Crook, O.M., Monti, M., Dezi, V., Willis, A.E. and Lilley, K.S. (2023) System-wide analysis of RNA and protein subcellular localization dynamics. Nature Methods. 1-12. doi:10.1038/s41592-023-02101-9

Christoforou, A., Mulvey, C.M., Breckels, L.M., Geladaki, A., Hurrell, T., Hayward, P.C., Naake, T., Gatto, L., Viner, R., Arias, A.M. and Lilley, K.S. (2016) A draft map of the mouse pluripotent stem cell spatial proteome. Nature Communications. 7 (1). doi:10.1038/ncomms9992

Barylyuk, K., Koreny, L., Ke, H., Butterworth, S., Crook, O.M., Lassadi, I., Gupta, V., Tromer, E., Mourier, T., Stevens, T.J., Breckels, L.M., Pain, A., Lilley, K.S. and Waller, R.F. (2020) A Comprehensive Subcellular Atlas of the Toxoplasma Proteome via hyperLOPIT Provides Spatial Context for Protein Functions. Cell Host and Microbe. 28 (5), 752-766.e9. doi:10.1016/j.chom.2020.09.011

Moloney, N.M., Barylyuk, K., Tromer, E., Crook, O.M., Breckels, L.M., Lilley, K.S., Waller, R.F. and MacGregor, P. (2023) Mapping diversity in African trypanosomes using high resolution spatial proteomics. Nature Communications. 14 (1), 4401. doi:10.1038/s41467-023-40125-z

Note: These markers are provided as a starting point to generate reliable sets of organelle markers but still need to be verified against any new data in the light of the quantitative data and the study conditions.

See also

Author

Laurent Gatto

Examples

pRolocmarkers()
#> 14 marker lists (version 2) available:
#> Arabidopsis thaliana [atha]:
#>  Ids: TAIR, 543 markers
#> Drosophila melanogaster [dmel]:
#>  Ids: Uniprot, 179 markers
#> Gallus gallus [ggal]:
#>  Ids: IPI, 102 markers
#> Homo sapiens [hsap]:
#>  Ids: Uniprot, 872 markers
#> Homo sapiens [hsap_christopher]:
#>  Ids: Uniprot, 1509 markers
#> Homo sapiens [hsap_geladaki]:
#>  Ids: Uniprot, 579 markers
#> Homo sapiens [hsap_itzhak]:
#>  Ids: Uniprot, 1076 markers
#> Homo sapiens [hsap_villaneuva]:
#>  Ids: Uniprot, 682 markers
#> Mus musculus [mmus]:
#>  Ids: Uniprot, 937 markers
#> Mus musculus [mmus_christoforou]:
#>  Ids: Uniprot, 922 markers
#> Saccharomyces cerevisiae [scer_sgd]:
#>  Ids: SGD, 259 markers
#> Saccharomyces cerevisiae [scer_uniprot]:
#>  Ids: Uniprot, 259 markers
#> Toxoplasma gondii [toxo_barylyuk]:
#>  Ids: ToxoDB gene identifier, 718 markers
#> Trypanosoma brucei [tryp_moloney]:
#>  Ids: TriTrypDB gene identifier, 891 markers
pRolocmarkers("hsap")
#>               P08865               P0CW22               P15880 
#>       "40S Ribosome"       "40S Ribosome"       "40S Ribosome" 
#>               P22090               P23396               P25398 
#>       "40S Ribosome"       "40S Ribosome"       "40S Ribosome" 
#>               P39019               P42677               P46781 
#>       "40S Ribosome"       "40S Ribosome"       "40S Ribosome" 
#>               P46782               P46783               P60866 
#>       "40S Ribosome"       "40S Ribosome"       "40S Ribosome" 
#>               P61247               P62081               P62241 
#>       "40S Ribosome"       "40S Ribosome"       "40S Ribosome" 
#>               P62244               P62249               P62263 
#>       "40S Ribosome"       "40S Ribosome"       "40S Ribosome" 
#>               P62266               P62269               P62273 
#>       "40S Ribosome"       "40S Ribosome"       "40S Ribosome" 
#>               P62277               P62280               P62701 
#>       "40S Ribosome"       "40S Ribosome"       "40S Ribosome" 
#>               P62753               P62841             P62847-2 
#>       "40S Ribosome"       "40S Ribosome"       "40S Ribosome" 
#>               P62851               P62854               P62857 
#>       "40S Ribosome"       "40S Ribosome"       "40S Ribosome" 
#>               P62861               P62979               P63220 
#>       "40S Ribosome"       "40S Ribosome"       "40S Ribosome" 
#>               P05386               P05387               P05388 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P18077               P18124               P18621 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P26373               P27635               P30050 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P32969               P35268               P36578 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P39023               P40429               P42766 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P46776               P46777               P46778 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P46779               P47914               P49207 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P50914               P61254               P61313 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P61353               P61513               P61927 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P62424               P62750               P62829 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P62888               P62899               P62906 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P62910             P62913-2               P62917 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P62987               P63173               P83731 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               P83881               P84098               Q02543 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               Q02878               Q07020               Q969Q0 
#>       "60S Ribosome"       "60S Ribosome"       "60S Ribosome" 
#>               Q9Y3U8               Q01518               P13796 
#>       "60S Ribosome" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>               P63261               P28289               O15143 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>               O15144               O15145               O15511 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>               O75369             O75369-8               P06753 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>             P06753-2             P06753-3             P06753-5 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>               P09493            P09493-10             P09493-5 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>               P12814               P23528               P47755 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>               P47756             P47756-2               P52907 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>               P59998               P60709             P60981-2 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>               P61158               P61160               P67936 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>               P68032               Q16658               Q562R1 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>               Q92747               Q9BPX5               Q9BR76 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>               Q9NYL9               Q9NZ32               Q9NZR1 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>             Q9P1U1-3               Q9Y281               Q9Y4G6 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>             P06753-6               Q14847               P07737 
#> "Actin Cytoskeleton" "Actin Cytoskeleton" "Actin Cytoskeleton" 
#>               Q9UJW0               P34932               P22102 
#> "Actin Cytoskeleton"            "Cytosol"            "Cytosol" 
#>               Q8TCU6               O60841               Q12882 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>             P49915-2               Q04446               Q6XQN6 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               O43847             Q06210-2               Q3KQV9 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               O14841               P19971               Q16543 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               P36871               Q9Y617               P60891 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               P16152               Q9HAB8               Q96KP4 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               P32119               Q9NR45               P19623 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               P53602               O43765               P29218 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               O95336               P16930             P08243-2 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               P34949               Q9HA64               P49593 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               P61081               P00492               Q9P2T1 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               Q15274               O95394               Q9H773 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               P29762               Q6IA69               Q14376 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               P09467               Q96AT9               Q13630 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               Q9H8S9               Q96BN8               P37837 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               P48637               Q9NR50               O75822 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               A0AVT1             P15170-2               P31153 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               Q04760               P50225               P18440 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               P51570               Q9BRA2               P50452 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               A6NDG6               Q9NRX4               Q9BQC3 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               O14732               Q9H2P9               O43175 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               Q9NT62               P52788               P07741 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               P49588               P09488               P49591 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               P49589               Q96C23               P49902 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>               Q7L5D6               P30711               Q9UPN7 
#>            "Cytosol"            "Cytosol"            "Cytosol" 
#>             Q9NYU2-2               Q5JRA6               P13667 
#>                 "ER"                 "ER"                 "ER" 
#>               P30101               P07237               P16615 
#>                 "ER"                 "ER"                 "ER" 
#>               P04843               P16435               P08240 
#>                 "ER"                 "ER"                 "ER" 
#>               O60568               P27824               Q12797 
#>                 "ER"                 "ER"                 "ER" 
#>               Q8NBJ5             Q15084-3               Q9UGP8 
#>                 "ER"                 "ER"                 "ER" 
#>               P48449             Q13438-4               Q02809 
#>                 "ER"                 "ER"                 "ER" 
#>               O95479             Q9H0X9-2               Q13724 
#>                 "ER"                 "ER"                 "ER" 
#>             O15320-8               Q14571               O75477 
#>                 "ER"                 "ER"                 "ER" 
#>               Q9BS26               Q7Z2K6             P14314-2 
#>                 "ER"                 "ER"                 "ER" 
#>               P30040               P50454               Q8NBS9 
#>                 "ER"                 "ER"                 "ER" 
#>               Q9Y4P3               Q9BPW9             Q9BZQ6-2 
#>                 "ER"                 "ER"                 "ER" 
#>               P07099               O15269               Q16850 
#>                 "ER"                 "ER"                 "ER" 
#>               O15270               Q14554             Q969N2-5 
#>                 "ER"                 "ER"                 "ER" 
#>               Q96JJ7             Q969V3-2               Q92611 
#>                 "ER"                 "ER"                 "ER" 
#>               Q6UWW8               Q6P1M0               P39656 
#>                 "ER"                 "ER"                 "ER" 
#>               Q9BT09               Q96S52               O00469 
#>                 "ER"                 "ER"                 "ER" 
#>               Q86UL3               Q15293               O95302 
#>                 "ER"                 "ER"                 "ER" 
#>               Q96DZ1               Q15005               Q643R3 
#>                 "ER"                 "ER"                 "ER" 
#>               Q9H3N1               Q8NBM4               O43292 
#>                 "ER"                 "ER"                 "ER" 
#>               Q9H488               O95881             Q9BU23-2 
#>                 "ER"                 "ER"                 "ER" 
#>               Q6IAN0               Q8TC12               Q9UBM7 
#>                 "ER"                 "ER"                 "ER" 
#>               P61619               Q9NZ01               P26885 
#>                 "ER"                 "ER"                 "ER" 
#>             Q9H6R6-2               Q99442               P60468 
#>                 "ER"                 "ER"                 "ER" 
#>               Q92604               O00400             Q86TM6-2 
#>                 "ER"                 "ER"                 "ER" 
#>               P61009               Q9P0I2             Q15011-3 
#>                 "ER"                 "ER"                 "ER" 
#>             Q9BV94-2               Q9UNW1               P43307 
#>                 "ER"                 "ER"                 "ER" 
#>             O75298-3               Q8N5M9               Q96HR9 
#>                 "ER"                 "ER"                 "ER" 
#>               Q96E22               O75845               Q9P2X0 
#>                 "ER"                 "ER"                 "ER" 
#>               Q6Y288               Q9BQB6               Q8IYK4 
#>                 "ER"                 "ER"                 "ER" 
#>               Q7Z4H8               O15460               O95470 
#>                 "ER"                 "ER"                 "ER" 
#>               Q2TAA5               Q8TCJ2               O60476 
#>                 "ER"                 "ER"              "Golgi" 
#>               O75063               P26572               Q10469 
#>              "Golgi"              "Golgi"              "Golgi" 
#>               Q10472               Q14789             Q14789-2 
#>              "Golgi"              "Golgi"              "Golgi" 
#>             Q14789-4               Q16706               Q5SRI9 
#>              "Golgi"              "Golgi"              "Golgi" 
#>               Q7LGA3               Q8TBA6             Q8TBA6-2 
#>              "Golgi"              "Golgi"              "Golgi" 
#>               Q9NX62               Q9NXS2               Q9NY97 
#>              "Golgi"              "Golgi"              "Golgi" 
#>             Q9NY97-2               Q9P2E5             Q495W5-2 
#>              "Golgi"              "Golgi"              "Golgi" 
#>               A6NKF9               Q9BYC5               O94766 
#>              "Golgi"              "Golgi"              "Golgi" 
#>               P49641               Q8NBZ7               Q8NEW0 
#>              "Golgi"              "Golgi"              "Golgi" 
#>               O14653             O14653-3               P22083 
#>              "Golgi"              "Golgi"              "Golgi" 
#>               Q13948               Q86SF2               P33908 
#>              "Golgi"              "Golgi"              "Golgi" 
#>               Q8NCL4               O00461               Q8N4A0 
#>              "Golgi"              "Golgi"              "Golgi" 
#>               O00115             O00115-2               O00462 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>               O00754             P04062-4               P04066 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>               P06865               P07339               P07602 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>               P07686               P07858               P09668 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>               P10253               P10619             P10619-2 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>               P15586               P17050               P34059 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>               P38571             P38571-2               P43234 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>               Q13571               Q14108               Q6ZP29 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>             Q6ZP29-2               Q8NBJ9               Q9NUN5 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>             Q9NUN5-3               Q9UBX1               Q9UHL4 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>             Q8WWB7-2               O14773               Q9Y646 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>               P35475               Q8NCC3             Q9HAT2-2 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>             Q02083-2               Q01459               Q96RQ9 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>             Q68CP4-2               P15289               P54802 
#>           "Lysosome"           "Lysosome"           "Lysosome" 
#>               P40939               Q13423               P43304 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>             O60313-2               Q99798               Q3SY69 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>             Q12931-2               Q02218               Q5JTZ9 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>             P28331-4               Q5JRX3               Q9Y6N5 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P49411               P00367               O94826 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               O95202               P09622               Q16822 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P31930               P23786               P31040 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P23368               P24752               P49448 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P05091             P56181-2               P22695 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               Q10713               O43615               P22033 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               Q6NUK1             P49821-2             Q02252-2 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P30837               P13804               P42765 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               O75390               P10515               P22570 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>             Q96CM8-3               P12694               P30084 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               O75306               P55809               O00411 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P50213             Q00325-2               O75439 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               O00330               P04181               P30405 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P26440             Q9Y305-2               P11182 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               Q53H12               O75947               O95299 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P53007               P22830               P16219 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               Q4G0N4               P13995               Q16134 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P30038               P36542               O96008 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P30042               P53597               Q02127 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               Q3ZCQ8               Q15118               P19404 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P47985               O75027               O43837 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               Q8N0X4               Q6NVY1               P46199 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P43897               P13073               P45954 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               O15382               P51649               P27144 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               O00217               P32322               P51553 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P42126               P08574               O43181 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>             Q15070-2               O75208               P62072 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P18859               O43716               O60830 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               O75251               P10606               O75964 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               O43678               O14925               P00403 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               O75380               Q9Y5L4             O95169-3 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               Q9Y5J9               P30049               P09669 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               Q9Y5J7               Q99766               P15954 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P10109               O95563               P23434 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               O95167               O43676             Q8IUX1-4 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               P0DJ07             O14521-2             O00142-2 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               Q9UKU7               Q8NI60               Q15031 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               Q9P0J1               Q9NUB1               O75127 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               O43819               P07919               O75879 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               O14874               Q16740               Q5HYK3 
#>      "Mitochondrion"      "Mitochondrion"      "Mitochondrion" 
#>               Q15120               Q5T160               P46013 
#>      "Mitochondrion"      "Mitochondrion"            "Nucleus" 
#>               P02545               Q14683               P20700 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               P11388               A6NHR9               Q9UQE7 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               O95347               Q9NTJ3               Q03252 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               O60264               Q9NRL2               Q9NTI5 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q5UIP0               Q29RF7               Q9Y5B9 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q9NR30               Q9GZR7               P24928 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q8N1F7               Q8NI27             P52948-5 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               O94776               Q92922               P11387 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q7LBC6               P16401               P42167 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               O95239               Q5SSJ5               Q9Y2U8 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               O60934               Q9H0A0               P10412 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               P43246               P16402             Q15424-4 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q08945               Q9NXV6               Q01826 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>             P51532-5               P16403               Q8NFC6 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               P23193               O60341             P42167-2 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q86YP4               P42695               Q96ST3 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               O14981               O00567             P35251-2 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q92522               Q8WXF1               Q8WXI9 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q15059               Q5QJE6               P49711 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               O75694               Q96FV9               O60216 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               P35249               P35269               P62805 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q9NX58               Q15554               Q14807 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>             Q969G3-2             O43684-2               Q13573 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               P85037               Q96SB8               P07305 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               O14647               P84243               P68431 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               P25490             Q03164-2               P35250 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q13769               Q9Y3T9             Q92925-3 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               P20585               Q9H8H0               P40937 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q13111               P40938               P18887 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>             Q12824-2               P13984               Q9H9Y6 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               O43818               O15525             Q9Y3Y2-4 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>             Q76FK4-4               O15047               Q13415 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               P0C0S5               Q8IXM2               Q13416 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q9UBD5             Q9P0W2-3               P52701 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               O75475               O60885               P26583 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               O15347               Q9H7Z6               Q9NVP1 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q9BVJ6               Q9BQ39               O15381 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q969X6               P55769               Q9NV31 
#>            "Nucleus"            "Nucleus"            "Nucleus" 
#>               Q9NPE3               O00116               O14832 
#>            "Nucleus"         "Peroxisome"         "Peroxisome" 
#>               O15254               O43808               O43933 
#>         "Peroxisome"         "Peroxisome"         "Peroxisome" 
#>               O75381             O75381-2               O96011 
#>         "Peroxisome"         "Peroxisome"         "Peroxisome" 
#>               P09110               P0C024               P28328 
#>         "Peroxisome"         "Peroxisome"         "Peroxisome" 
#>               P51659               Q08426               Q13608 
#>         "Peroxisome"         "Peroxisome"         "Peroxisome" 
#>               Q15067             Q15067-2               Q7Z412 
#>         "Peroxisome"         "Peroxisome"         "Peroxisome" 
#>             Q7Z412-2               Q86WA8               Q8WVX9 
#>         "Peroxisome"         "Peroxisome"         "Peroxisome" 
#>               Q9BY49               Q9NR77               Q9NUI1 
#>         "Peroxisome"         "Peroxisome"         "Peroxisome" 
#>               Q9P0Z9               Q9UKG9               Q9Y6I8 
#>         "Peroxisome"         "Peroxisome"         "Peroxisome" 
#>               O14734               O00161               O14786 
#>         "Peroxisome"                 "PM"                 "PM" 
#>             O14786-3             O14828-2               O14910 
#>                 "PM"                 "PM"                 "PM" 
#>               O15031               O15394               O43760 
#>                 "PM"                 "PM"                 "PM" 
#>             O60241-4               O60449             O60449-3 
#>                 "PM"                 "PM"                 "PM" 
#>             O60462-4             O60637-3             O75019-2 
#>                 "PM"                 "PM"                 "PM" 
#>               O75955               O75976               O94856 
#>                 "PM"                 "PM"                 "PM" 
#>             O94910-2             O95297-2               O95502 
#>                 "PM"                 "PM"                 "PM" 
#>               P01892               P04216               P05023 
#>                 "PM"                 "PM"                 "PM" 
#>               P05026               P05106               P05362 
#>                 "PM"                 "PM"                 "PM" 
#>               P05556             P06213-2             P06756-3 
#>                 "PM"                 "PM"                 "PM" 
#>               P07204               P07949             P08195-2 
#>                 "PM"                 "PM"                 "PM" 
#>               P08571             P08575-2               P08582 
#>                 "PM"                 "PM"                 "PM" 
#>               P08648               P08754               P09543 
#>                 "PM"                 "PM"                 "PM" 
#>             P10586-2               P11166               P11215 
#>                 "PM"                 "PM"                 "PM" 
#>             P12318-2               P13164               P13591 
#>                 "PM"                 "PM"                 "PM" 
#>             P13591-1               P13598               P13612 
#>                 "PM"                 "PM"                 "PM" 
#>             P14209-3               P14415            P16070-18 
#>                 "PM"                 "PM"                 "PM" 
#>               P16150             P16284-3               P17301 
#>                 "PM"                 "PM"                 "PM" 
#>               P17677             P17813-2               P18084 
#>                 "PM"                 "PM"                 "PM" 
#>               P19022             P19022-2               P19397 
#>                 "PM"                 "PM"                 "PM" 
#>               P20020             P20020-6               P20138 
#>                 "PM"                 "PM"                 "PM" 
#>               P20701               P20702               P21589 
#>                 "PM"                 "PM"                 "PM" 
#>               P22794               P23229             P23634-4 
#>                 "PM"                 "PM"                 "PM" 
#>               P26010               P26992               P27105 
#>                 "PM"                 "PM"                 "PM" 
#>             P27701-2             P29323-2               P32004 
#>                 "PM"                 "PM"                 "PM" 
#>               P32249               P32942               P32970 
#>                 "PM"                 "PM"                 "PM" 
#>             P33527-4               P35222               P35408 
#>                 "PM"                 "PM"                 "PM" 
#>             P35613-2               P36383               P38570 
#>                 "PM"                 "PM"                 "PM" 
#>             P41597-2               P41732             P42857-2 
#>                 "PM"                 "PM"                 "PM" 
#>             P43250-2               P46939               P48509 
#>                 "PM"                 "PM"                 "PM" 
#>             P48960-2               P50895               P51674 
#>                 "PM"                 "PM"                 "PM" 
#>               P54762             P55196-3               P63092 
#>                 "PM"                 "PM"                 "PM" 
#>             P78552-2               P84095             P98155-2 
#>                 "PM"                 "PM"                 "PM" 
#>               P98172               Q01650               Q01814 
#>                 "PM"                 "PM"                 "PM" 
#>               Q03405               Q04941               Q08722 
#>                 "PM"                 "PM"                 "PM" 
#>               Q12846               Q13449               Q13491 
#>                 "PM"                 "PM"                 "PM" 
#>             Q13740-2               Q14254               Q14699 
#>                 "PM"                 "PM"                 "PM" 
#>             Q14982-2               Q15262               Q15722 
#>                 "PM"                 "PM"                 "PM" 
#>               Q16720             Q6GTX8-3               Q6IA17 
#>                 "PM"                 "PM"                 "PM" 
#>               Q6UXK5               Q6X4W1               Q7Z2K8 
#>                 "PM"                 "PM"                 "PM" 
#>               Q7Z3B1             Q7Z403-2               Q7Z6M3 
#>                 "PM"                 "PM"                 "PM" 
#>             Q8IWK6-3               Q8IX19               Q8IYJ0 
#>                 "PM"                 "PM"                 "PM" 
#>               Q8N0W4               Q8N2Q7             Q8N8Q9-2 
#>                 "PM"                 "PM"                 "PM" 
#>               Q8N9M5             Q8NC67-3               Q8NFZ4 
#>                 "PM"                 "PM"                 "PM" 
#>             Q8NHJ6-2             Q8NHJ6-3               Q8NHL6 
#>                 "PM"                 "PM"                 "PM" 
#>               Q8TBP5             Q8TCZ2-6               Q92692 
#>                 "PM"                 "PM"                 "PM" 
#>             Q92823-3               Q92854             Q92859-2 
#>                 "PM"                 "PM"                 "PM" 
#>             Q96D96-4             Q96F46-2               Q96PE1 
#>                 "PM"                 "PM"                 "PM" 
#>               Q96S97               Q99569             Q99569-2 
#>                 "PM"                 "PM"                 "PM" 
#>               Q99571             Q99572-8               Q99795 
#>                 "PM"                 "PM"                 "PM" 
#>             Q9BY67-5               Q9GZY6               Q9H2W1 
#>                 "PM"                 "PM"                 "PM" 
#>               Q9H6B4             Q9H6X2-5               Q9H7M9 
#>                 "PM"                 "PM"                 "PM" 
#>               Q9H813               Q9HAR2               Q9NPR9 
#>                 "PM"                 "PM"                 "PM" 
#>               Q9NQS5             Q9NT68-2               Q9NW97 
#>                 "PM"                 "PM"                 "PM" 
#>               Q9NZ94               Q9P121               Q9UBG0 
#>                 "PM"                 "PM"                 "PM" 
#>             Q9UHW9-3               Q9UIW2             Q9Y219-2 
#>                 "PM"                 "PM"                 "PM" 
#>             Q9Y287-2             Q9Y2J2-2               Q9Y624 
#>                 "PM"                 "PM"                 "PM" 
#>               Q9Y639               Q9UIQ6               P78536 
#>                 "PM"                 "PM"                 "PM" 
#>               P08311               Q9NWQ8               P04839 
#>                 "PM"                 "PM"                 "PM" 
#>               P14384               P11234               P07333 
#>                 "PM"                 "PM"                 "PM" 
#>               Q02487               P63218               P30825 
#>                 "PM"                 "PM"                 "PM" 
#>               Q9NRM0               P41231               P32246 
#>                 "PM"                 "PM"                 "PM" 
#>               Q9NPG4               Q13433               P98153 
#>                 "PM"                 "PM"                 "PM" 
#>               P51511               P31641               Q99460 
#>                 "PM"                 "PM"         "Proteasome" 
#>               Q13200               O43242               P25786 
#>         "Proteasome"         "Proteasome"         "Proteasome" 
#>               O00232               Q15008               O00231 
#>         "Proteasome"         "Proteasome"         "Proteasome" 
#>               Q06323               O14818               Q9UNM6 
#>         "Proteasome"         "Proteasome"         "Proteasome" 
#>               P60900               P55036               P25789 
#>         "Proteasome"         "Proteasome"         "Proteasome" 
#>               Q9UL46               P20618             P25788-2 
#>         "Proteasome"         "Proteasome"         "Proteasome" 
#>               P28062               P61289               P25787 
#>         "Proteasome"         "Proteasome"         "Proteasome" 
#>               P28066               P49720               P49721 
#>         "Proteasome"         "Proteasome"         "Proteasome" 
#>               P51665               P40306               O75832 
#>         "Proteasome"         "Proteasome"         "Proteasome" 
#>               P28070               P28072               P48556 
#>         "Proteasome"         "Proteasome"         "Proteasome" 
#>               P28065               O00487             O95456-2 
#>         "Proteasome"         "Proteasome"         "Proteasome" 
#>               Q99436             P62195-2               P62191 
#>         "Proteasome"         "Proteasome"         "Proteasome" 
#>               P43686               P35998 
#>         "Proteasome"         "Proteasome" 
table(pRolocmarkers("hsap"))
#> 
#>       40S Ribosome       60S Ribosome Actin Cytoskeleton            Cytosol 
#>                 33                 46                 45                 77 
#>                 ER              Golgi           Lysosome      Mitochondrion 
#>                 92                 34                 42                134 
#>            Nucleus                 PM         Peroxisome         Proteasome 
#>                116                190                 27                 36 

## Old markers
pRolocmarkers("hsap", version = "2")["Q9BPW9"]
#> Q9BPW9 
#>   "ER" 
pRolocmarkers("hsap", version = "1")["Q9BPW9"]
#>                  Q9BPW9 
#> "Endoplasmic Reticulum"