Organelle markers

This function retrieves a list of organelle markers or, if no species is provided, prints a description of available marker sets. The markers can be added to and MSnSet using the addMarkers function.

pRolocmarkers(species)

Arguments

species

The species of interest.

Value

Prints a description of the available marker lists if species is missing or a named character with organelle markers.

Details

The markers have been contributed by various members of the Cambridge Centre for Proteomics, in particular Dr Dan Nightingale for yeast, Dr Andy Christoforou and Dr Claire Mulvey for human, Dr Arnoud Groen for Arabodopsis and Dr Claire Mulvey for mouse. In addition, original (curated) markers from the pRolocdata datasets have been extracted (see pRolocdata for details and references). Curation involved verification of publicly available subcellular localisation annotation based on the curators knowledge of the organelles/proteins considered and tracing the original statement in the literature.

These markers are provided as a starting point to generate reliable sets of organelle markers but still need to be verified against any new data in the light of the quantitative data and the study conditions.

See also

addMarkers to add markers to an MSnSet and markers for more information about marker encoding.

Examples

pRolocmarkers()

#> 7 marker lists available:
#> Arabidopsis thaliana [atha]:
#>  Ids: TAIR, 543 markers
#> Drosophila melanogaster [dmel]:
#>  Ids: Uniprot, 179 markers
#> Gallus gallus [ggal]:
#>  Ids: IPI, 102 markers
#> Homo sapiens [hsap]:
#>  Ids: Uniprot, 872 markers
#> Mus musculus [mmus]:
#>  Ids: Uniprot, 937 markers
#> Saccharomyces cerevisiae [scer_sgd]:
#>  Ids: SGD, 259 markers
#> Saccharomyces cerevisiae [scer_uniprot]:
#>  Ids: Uniprot Accession, 259 markers

table(pRolocmarkers("atha"))

#> 
#>           ENV            ER      ER lumen   ER membrane         Golgi 
#>            46            30             9            42            27 
#> Mitochondrion       Plastid            PM      Ribosome           STR 
#>           103            48            60             8            83 
#>           TGN           THY       Vacuole 
#>            13            56            18 

table(pRolocmarkers("hsap"))

#> 
#>          40S Ribosome          60S Ribosome    actin cytoskeleton 
#>                    33                    46                    45 
#>               Cytosol Endoplasmic Reticulum       Golgi Apparatus 
#>                    77                    92                    34 
#>              Lysosome          Mitochondria               Nucleus 
#>                    42                   134                   116 
#>            PEROXISOME            plasma mem            Proteasome 
#>                    27                   190                    36