PerTurbo parameter optimisation

Classification parameter optimisation for the PerTurbo algorithm

perTurboOptimisation(object, fcol = "markers", pRegul = 10^(seq(from =
  -1, to = 0, by = 0.2)), sigma = 10^(seq(from = -1, to = 1, by = 0.5)),
  inv = c("Inversion Cholesky", "Moore Penrose", "solve", "svd"),
  reg = c("tikhonov", "none", "trunc"), times = 1, test.size = 0.2,
  xval = 5, fun = mean, seed, verbose = TRUE)

Arguments

object
An instance of class "MSnSet".

fcol
The feature meta-data containing marker definitions. Default is markers.

pRegul
The hyper-parameter for the regularisation (values are in ]0,1] ). If reg =="trunc", pRegul is for the percentage of eigen values in matrix. If reg =="tikhonov", then 'pRegul' is the parameter for the tikhonov regularisation. Available configurations are : "Inversion Cholesky" - ("tikhonov" / "none"), "Moore Penrose" - ("tikhonov" / "none"), "solve" - ("tikhonov" / "none"), "svd" - ("tikhonov" / "none" / "trunc").

sigma
The hyper-parameter.

inv
The type of algorithm used to invert the matrix. Values are : "Inversion Cholesky" (chol2inv), "Moore Penrose" (ginv), "solve" (solve), "svd" (svd). Default value is "Inversion Cholesky".

reg
The type of regularisation of matrix. Values are "none", "trunc" or "tikhonov". Default value is "tikhonov".

times
The number of times internal cross-validation is performed. Default is 100.

test.size
The size of test data. Default is 0.2 (20 percent).

xval
The n-cross validation. Default is 5.

fun
The function used to summarise the times macro F1 matrices.

seed
The optional random number generator seed.

verbose
A logical defining whether a progress bar is displayed.

Value

An instance of class "GenRegRes".

Details

Note that when performance scores precision, recall and (macro) F1 are calculated, any NA values are replaced by 0. This decision is motivated by the fact that any class that would have either a NA precision or recall would result in an NA F1 score and, eventually, a NA macro F1 (i.e. mean(F1)). Replacing NAs by 0s leads to F1 values of 0 and a reduced yet defined final macro F1 score.

See also