plsda classification

Classification using the partial least square distcriminant analysis algorithm.

plsdaClassification(object, assessRes, scores = c("prediction", "all",
  "none"), ncomp, fcol = "markers", ...)

Arguments

object

An instance of class "MSnSet".

assessRes

An instance of class "GenRegRes", as generated by plsdaOptimisation.

scores

One of "prediction", "all" or "none" to report the score for the predicted class only, for all cluster or none.

ncomp

If assessRes is missing, a ncomp must be provided.

fcol

The feature meta-data containing marker definitions. Default is markers.

...

Additional parameters passed to plsda from package caret.

Value

An instance of class "MSnSet" with plsda and plsda.scores feature variables storing the classification results and scores respectively.

Examples

## not running this one for time considerations
library(pRolocdata)
data(dunkley2006)
## reducing parameter search space and iterations 
params <- plsdaOptimisation(dunkley2006, ncomp = c(3, 10),  times = 2)

#>   |                                                                              |                                                                      |   0%

#> Error in requireNamespaceQuietStop("klaR"): package klaR is required

params

#> standardGeneric for "params" defined from package "S4Vectors"
#> 
#> function (x, ...) 
#> standardGeneric("params")
#> <environment: 0x7f09b30>
#> Methods may be defined for arguments: x
#> Use  showMethods("params")  for currently available ones.

plot(params)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("numeric"), structure(function (x, ...) standardGeneric("params"), generic = structure("params", package = "S4Vectors"), package = "S4Vectors", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'params' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘params’ for signature ‘"numeric"’

f1Count(params)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (object,     ...) standardGeneric("f1Count"), generic = structure("f1Count", package = "pRoloc"), package = "pRoloc", group = list(), valueClass = character(0), signature = "object", default = `\001NULL\001`, skeleton = (function (object,     ...) stop("invalid call in method dispatch to 'f1Count' (no default method)",     domain = NA))(object, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘f1Count’ for signature ‘"standardGeneric"’

levelPlot(params)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (object,     ...) standardGeneric("levelPlot"), generic = structure("levelPlot", package = "pRoloc"), package = "pRoloc", group = list(), valueClass = character(0), signature = "object", default = `\001NULL\001`, skeleton = (function (object,     ...) stop("invalid call in method dispatch to 'levelPlot' (no default method)",     domain = NA))(object, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘levelPlot’ for signature ‘"standardGeneric"’

getParams(params)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (object,     ...) standardGeneric("getParams"), generic = structure("getParams", package = "pRoloc"), package = "pRoloc", group = list(), valueClass = character(0), signature = "object", default = `\001NULL\001`, skeleton = (function (object,     ...) stop("invalid call in method dispatch to 'getParams' (no default method)",     domain = NA))(object, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘getParams’ for signature ‘"standardGeneric"’

res <- plsdaClassification(dunkley2006, params)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (object,     ...) standardGeneric("getParams"), generic = structure("getParams", package = "pRoloc"), package = "pRoloc", group = list(), valueClass = character(0), signature = "object", default = `\001NULL\001`, skeleton = (function (object,     ...) stop("invalid call in method dispatch to 'getParams' (no default method)",     domain = NA))(object, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘getParams’ for signature ‘"standardGeneric"’

getPredictions(res, fcol = "plsda")

#> Error in fData(object): object 'res' not found

getPredictions(res, fcol = "plsda", t = 0.9)

#> Error in fData(object): object 'res' not found

plot2D(res, fcol = "plsda")

#> Error in plot2D(res, fcol = "plsda"): object 'res' not found