Draw matrix of thetas to test

The possible weights to be considered is a sequence from 0 (favour auxiliary data) to 1 (favour primary data). Each possible combination of weights for nclass classes must be tested. The thetas function produces a weight matrix for nclass columns (one for each class) with all possible weight combinations (number of rows).

Usage

thetas(nclass, by = 0.5, length.out, verbose = TRUE)

Arguments

nclass

Number of marker classes

by

The increment of the weights. One of 1, 0.5, 0.25, 2, 0.1 or 0.05.

length.out

The desired length of the weight sequence.

verbose

A logical indicating if the weight sequences should be printed out. Default is TRUE.

Value

A matrix with all possible theta weight combinations.

Author

Lisa Breckels

Examples

dim(thetas(4, by = 0.5))
#> Weigths:
#>   (0, 0.5, 1)
#> [1] 81  4
dim(thetas(4, by = 0.2))
#> Weigths:
#>   (0, 0.2, 0.4, 0.6, 0.8, 1)
#> [1] 1296    4
dim(thetas(5, by = 0.2))
#> Weigths:
#>   (0, 0.2, 0.4, 0.6, 0.8, 1)
#> [1] 7776    5
dim(thetas(5, length.out = 5))
#> Weigths:
#>   (0, 0.25, 0.5, 0.75, 1)
#> [1] 3125    5
dim(thetas(6, by = 0.2))
#> Weigths:
#>   (0, 0.2, 0.4, 0.6, 0.8, 1)
#> [1] 46656     6